Fix warning
Introduced in commit 2efe70da01edd0fe4aa7dfab63da37147fdc9ee1.

+++ src/Utils/TreePropSheet/TreePropSheet.cpp
@@ -406,7 +406,7 @@ void CTreePropSheet::RefillPageTree()
 			HTREEITEM hChild = m_pwndPageTree->GetChildItem(hParent);
 			while (hChild)
 			{
-				if (m_pwndPageTree->GetItemData(hChild) == parentId)
+				if (static_cast<int>(m_pwndPageTree->GetItemData(hChild)) == parentId)
 				{
 					hItem = hChild;
 					break;
